refactor: prepare generic download page
